[WIP] Continuing refactoring of the CLI by moving database utilities and Google analytics utilities to the 'utils' folder.
